training/requirements.txt
torch>=2.0.0
onnx>=1.15.0
onnxruntime>=1.17.0
onnxruntime-tools>=1.7.0
onnxscript>=0.6.0
numpy>=1.24.0
librosa>=0.10.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
soundfile>=0.12.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
